update imports path for react and vanilla
