+++ spark/manifests/angular/package-angular.json
