+++ spark/manifests/react/package-react.json
