+++ spark/manifests/vanilla/package.json
