+++ spark/manifests/vanilla/spark-exports-react.js
@@ -1,9 +1,9 @@
-export { default as SprkAlert } from './components/alerts/react/SprkAlert';
-export { default as SprkButton } from './components/buttons/react/SprkButton';
+export { default as SprkAlert } from '../../components/alerts/react/SprkAlert';
+export { default as SprkButton } from '../../components/buttons/react/SprkButton';
 export {
   default as SprkDatePickerInput,
-} from './base/inputs/react/SprkDatePickerInput';
-export { default as SprkDictionary } from './components/dictionary/react/SprkDictionary';
+} from '../../base/inputs/react/SprkDatePickerInput';
+export { default as SprkDictionary } from '../../components/dictionary/react/SprkDictionary';
 export { default as SprkDivider } from './components/divider/react/SprkDivider/SprkDivider';
 export { default as SprkDropdown } from './components/dropdown/react/SprkDropdown/SprkDropdown';
 export {
@@ -12,24 +12,24 @@ export {
 export { default as SprkIcon } from './components/icons/react/SprkIcon/SprkIcon';
 export {
   default as SprkTextInput,
-} from './base/inputs/react/SprkTextInput';
+} from '../../base/inputs/react/SprkTextInput';
 export {
   default as SprkRevealInput,
-} from './base/inputs/react/SprkRevealInput';
+} from '../../base/inputs/react/SprkRevealInput';
 export {
   default as SprkSelectionInput,
-} from './base/inputs/react/SprkSelectionInput';
-export { default as SprkLink } from './base/links/react/SprkLink';
+} from '../../base/inputs/react/SprkSelectionInput';
+export { default as SprkLink } from '../../base/links/react/SprkLink';
 export { default as SprkTable } from './base/table/react/SprkTable';
-export { default as SprkMasthead } from './components/masthead/react/SprkMasthead';
-export { default as SprkPagination } from './components/pagination/react/SprkPagination';
-export { default as SprkPromo } from './components/promo/react/SprkPromo';
+export { default as SprkMasthead } from '../../components/masthead/react/SprkMasthead';
+export { default as SprkPagination } from '../../components/pagination/react/SprkPagination';
+export { default as SprkPromo } from '../../components/promo/react/SprkPromo';
 export { default as SprkSpinner } from './components/spinner/react/SprkSpinner';
-export { default as SprkList } from './base/lists/react/SprkList';
+export { default as SprkList } from '../../base/lists/react/SprkList';
 export {
   default as SprkListItem,
-} from './base/lists/react/SprkListItem';
-export { default as SprkTabs } from './components/tabs/react/SprkTabs';
+} from '../../base/lists/react/SprkListItem';
+export { default as SprkTabs } from '../../components/tabs/react/SprkTabs';
 export {
   default as SprkTabsPanel,
 } from './components/tabs/react/SprkTabsPanel';
@@ -44,5 +44,5 @@ export {
   default as SprkStackItem,
 } from './objects/stack/react/SprkStackItem';
 export { default as SprkAward } from './components/award/react/SprkAward';
-export { default as SprkCard } from './components/card/react/SprkCard';
+export { default as SprkCard } from '../../components/card/react/SprkCard';
 export { default as SprkModal } from './components/modal/react/SprkModal';
