--- spark/spark-polyfills.js
@@ -1,8 +1,0 @@
-// Polyfills
-import 'promise-polyfill/src/polyfill';
-import './utilities/polyfills/vanilla/ArrayFrom';
-import './utilities/polyfills/vanilla/StringIncludes';
-import './utilities/polyfills/vanilla/ArrayFind';
-import './utilities/polyfills/vanilla/NodeListForEach';
-import './utilities/polyfills/vanilla/classListSVG';
-import './utilities/polyfills/vanilla/ObjectPrepend';
